Reverting to previous commit.
Because I don't think floating point errors were really the source of
the synchronization errors I was trying to address.

+++ src/geom/Point.java
@@ -1,9 +1,9 @@
 package geom;
 
 public class Point {
-	public double x, y;
+	public float x, y;
 	
-	public Point(double x, double y) {
+	public Point(float x, float y) {
 		this.x = x;
 		this.y = y;
 	}
